backend/identity-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9
python-jose[cryptography]
passlib[bcrypt]
pika>=1.3.0
python-multipart
requests
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
